Added Access Control for Payment API

+++ HealthCare_Payment/src/com/PayService.java
@@ -6,6 +6,8 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 //For JSON
 import com.google.gson.*;
+import com.AccessFilter;
+
 //For XML
 import org.jsoup.*;
 import org.jsoup.parser.*;
@@ -19,55 +21,73 @@ public class PayService {
 	@GET
 	@Path("/")
 	@Produces(MediaType.TEXT_HTML)
-	public String readPaymentDetails() {
-		return paymentObj.readPaymentDetails();
+	public String readPaymentDetails(@HeaderParam("authString") String authString) {
+
+		if (AccessFilter.checkAccess(new String[] { "patient", "admin" }, authString)) {
+			return paymentObj.readPaymentDetails();
+		} else {
+			return null;
+		}
 	}
 
-	
 	@POST
 	@Path("/")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	@Produces(MediaType.TEXT_PLAIN)
 	public String insertPaymentDetails(@FormParam("paymentId") String paymentId,
 			@FormParam("paidAmount") String paidAmount, @FormParam("paymentDescription") String paymentDescription,
-			@FormParam("Hid") String Hid, @FormParam("userId") String userId) {
-		String output = paymentObj.insertPaymentDetails(paymentId, paidAmount, paymentDescription, Hid, userId);
-		return output;
+			@FormParam("Hid") String Hid, @FormParam("userId") String userId,
+			@HeaderParam("authString") String authString) {
+
+		if (AccessFilter.checkAccess(new String[] { "patient", "admin" }, authString)) {
+			String output = paymentObj.insertPaymentDetails(paymentId, paidAmount, paymentDescription, Hid, userId);
+			return output;
+		} else {
+			return null;
+		}
 	}
 
 	@PUT
 	@Path("/")
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.TEXT_PLAIN)
-	public String updatePaymentDetails(String paymentData) {
+	public String updatePaymentDetails(String paymentData, @HeaderParam("authString") String authString) {
 
-		// Convert the input string to a JSON object
-		JsonObject paymentObject = new JsonParser().parse(paymentData).getAsJsonObject();
+		if (AccessFilter.checkAccess(new String[] { "patient", "admin" }, authString)) {
+			// Convert the input string to a JSON object
+			JsonObject paymentObject = new JsonParser().parse(paymentData).getAsJsonObject();
 
-		// Read the values from the JSON object
-		String paymentId = paymentObject.get("paymentId").getAsString();
-		String paidAmount = paymentObject.get("paidAmount").getAsString();
-		String paymentDescription = paymentObject.get("paymentDescription").getAsString();
-		String Hid = paymentObject.get("Hid").getAsString();
-		String userId = paymentObject.get("userId").getAsString();
+			// Read the values from the JSON object
+			String paymentId = paymentObject.get("paymentId").getAsString();
+			String paidAmount = paymentObject.get("paidAmount").getAsString();
+			String paymentDescription = paymentObject.get("paymentDescription").getAsString();
+			String Hid = paymentObject.get("Hid").getAsString();
+			String userId = paymentObject.get("userId").getAsString();
 
-		String output = paymentObj.updatePaymentDetails(paymentId, paidAmount, paymentDescription, Hid, userId);
-		return output;
+			String output = paymentObj.updatePaymentDetails(paymentId, paidAmount, paymentDescription, Hid, userId);
+			return output;
+		} else {
+			return null;
+		}
 	}
 
 	@DELETE
 	@Path("/")
 	@Consumes(MediaType.APPLICATION_XML)
 	@Produces(MediaType.TEXT_PLAIN)
-	public String deletePaymentDetails(String paymentData) {
+	public String deletePaymentDetails(String paymentData, @HeaderParam("authString") String authString) {
 
-		// Convert the input string to an XML document
-		Document doc = Jsoup.parse(paymentData, "", Parser.xmlParser());
+		if (AccessFilter.checkAccess(new String[] { "patient", "admin" }, authString)) {
+			// Convert the input string to an XML document
+			Document doc = Jsoup.parse(paymentData, "", Parser.xmlParser());
 
-		// Read the value from the element <itemID>
-		String paymentId = doc.select("paymentId").text();
-		String output = paymentObj.deletePaymentDetails(paymentId);
-		return output;
+			// Read the value from the element <itemID>
+			String paymentId = doc.select("paymentId").text();
+			String output = paymentObj.deletePaymentDetails(paymentId);
+			return output;
+		} else {
+			return null;
+		}
 	}
 
 }
